Confirm dialog to delete reservations

+++ app/ngLibrary/src/app/admin/admin-home/admin-home.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { UsersService } from '../../services/users/users.service';
 import { CubiclesService } from '../../services/cubicles/cubicles.service';
 import { CareersService } from '../../services/careers/careers.service';
@@ -6,6 +6,8 @@ import { DepartmentsService } from '../../services/departments/departments.servi
 import { EmployeesService } from '../../services/employees/employees.service';
 import { ExternalUserService } from '../../services/externalUser/external-user.service';
 import { ReservationsService } from '../../services/reservations/reservations.service';
+import { NguiPopupComponent, NguiMessagePopupComponent } from '@ngui/popup';
+import { PopupConfirmComponent } from '../../home/home-dialogs/popup-confirm/popup-confirm.component';
 
 @Component({
   selector: 'app-admin-home',
@@ -14,6 +16,8 @@ import { ReservationsService } from '../../services/reservations/reservations.se
 })
 export class AdminHomeComponent implements OnInit {
 
+  @ViewChild(NguiPopupComponent) popup: NguiPopupComponent;
+  message: Boolean
   totalUsers: number
   totalCubicles: number
   totalCareers: number
@@ -29,7 +33,7 @@ export class AdminHomeComponent implements OnInit {
     private employeesService: EmployeesService,
     private externalUserService: ExternalUserService,
     private reservationsService: ReservationsService
-  ) { }
+  ) { this.message = false }
 
   ngOnInit() {
     this.usersService.getAll().then(data => {
@@ -52,10 +56,19 @@ export class AdminHomeComponent implements OnInit {
     })
   }
 
-  removeAll() {
-    this.reservationsService.removeAll().then(response => {
-      console.log(response)
-    }).catch(err => console.log(err))
-  }
+  openPopup() {
+      this.popup.open(PopupConfirmComponent, {
+        classNames: 'custom'
+        // closeButton: true
+      })
+    }
 
+  removeAll() {
+    this.openPopup()
+    if (this.message) {
+      this.reservationsService.removeAll().then(response => {
+      console.log(response)
+      }).catch(err => console.log(err))
+    }
+  }
 }
